Upgrades and refactoring
