+++ package.json
@@ -11,10 +11,10 @@
     "@use-gesture/react": "^10.2.15",
     "html-to-image": "^1.9.0",
     "immer": "^9.0.15",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
-    "react-spring": "^9.4.5",
+    "react-spring": "^9.5.5",
     "react-use-measure": "^2.1.1",
     "use-onclickoutside": "^0.4.1",
     "uuid": "^8.3.2",
@@ -30,7 +30,7 @@
     "postcss": "^8.4.14",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
-    "tailwindcss": "^3.1.2",
+    "tailwindcss": "^3.2.0",
     "typescript": "^4.8.4",
     "vite": "^2.9.9"
   }
