+++ src/store/chordChart/chordChart.store.ts
@@ -1,9 +1,9 @@
 import create from 'zustand'
 import { devtools, persist } from 'zustand/middleware'
-import { createConfigSlice, ConfigSlice } from './config.slice'
-import { createChordLabelSlice, ChordLabelSlice } from './chordLabel.slice'
-import { createGridSlice, GridSlice } from './grid.slice'
-import { createNotesSlice, NotesSlice } from './notes.slice'
+import { createConfigSlice, ConfigSlice } from './slices/config.slice'
+import { createChordLabelSlice, ChordLabelSlice } from './slices/chordLabel.slice'
+import { createGridSlice, GridSlice } from './slices/grid.slice'
+import { createNotesSlice, NotesSlice } from './slices/notes.slice'
 
 export type State = ConfigSlice & ChordLabelSlice & GridSlice & NotesSlice
 export type Middlewares = [['zustand/devtools', never], ['zustand/persist', unknown]] //[['zustand/devtools', never]];
