+++ src/store/chordChart/slices/config.slice.ts
@@ -1,5 +1,5 @@
 import { StateCreator } from 'zustand'
-import { State, Middlewares } from './chordChart.store'
+import { State, Middlewares } from '../chordChart.store'
 import produce from 'immer'
 
 export type ConfigSlice = {
